Separate pages from components
