+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import { BrowserRouter as Router } from "react-router-dom";
 import { ThemeProvider, CSSReset } from "@chakra-ui/core";
 
-import App from "./components/app";
+import App from "./pages/app";
 
 ReactDOM.render(
   <React.StrictMode>
