+++ src/pages/launch-pad.js
@@ -18,8 +18,8 @@ import {
 } from "@chakra-ui/core";
 
 import { useSpaceX } from "../utils/use-space-x";
-import Error from "./error";
-import Breadcrumbs from "./breadcrumbs";
+import Error from "../components/error";
+import Breadcrumbs from "../components/breadcrumbs";
 import { LaunchItem } from "./launches";
 
 export default function LaunchPad() {
