+++ src/pages/launch.js
@@ -24,8 +24,8 @@ import {
 
 import { useSpaceX } from "../utils/use-space-x";
 import { formatDateTime } from "../utils/format-date";
-import Error from "./error";
-import Breadcrumbs from "./breadcrumbs";
+import Error from "../components/error";
+import Breadcrumbs from "../components/breadcrumbs";
 
 export default function Launch() {
   let { launchId } = useParams();
